requirements.txt
Flask==1.1.2
google-cloud-pubsub==1.4.2
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
